net_buf: change return type of net_buf_max_len as size_t
Change net_buf_max_len to use size_t as return value type to align with
the rest of the net_buf APIs.

+++ lib/net_buf/buf_simple.c
@@ -620,7 +620,7 @@ size_t net_buf_simple_tailroom(const struct net_buf_simple *buf)
 	return buf->size - net_buf_simple_headroom(buf) - buf->len;
 }
 
-uint16_t net_buf_simple_max_len(const struct net_buf_simple *buf)
+size_t net_buf_simple_max_len(const struct net_buf_simple *buf)
 {
 	return buf->size - net_buf_simple_headroom(buf);
 }
